static_cast UMA 'enum class's to int
Otherwise, the Chromium expansions complain about not being able to add
and needing explicit conversions.

R=mark@chromium.org
BUG=crashpad:100

+++ util/misc/metrics.cc
@@ -29,8 +29,9 @@ void Metrics::CrashReportSize(FileHandle file) {
 
 // static
 void Metrics::ExceptionCaptureResult(CaptureResult result) {
-  UMA_HISTOGRAM_ENUMERATION(
-      "Crashpad.ExceptionCaptureResult", result, CaptureResult::kMaxValue);
+  UMA_HISTOGRAM_ENUMERATION("Crashpad.ExceptionCaptureResult",
+                            static_cast<int32_t>(result),
+                            static_cast<int32_t>(CaptureResult::kMaxValue));
 }
 
 // static
